Type Player_80036978's output parameter as Vec3*
The parameter was declared s32, so the caller cast &world_pos to s32 and the body cast it back to Vec3* under an Eliminate-cast todo. Both translation units compile to byte-identical objects.

+++ src/melee/if/ifmagnify.c
@@ -296,7 +296,7 @@ void ifMagnify_802FBBDC(HSD_GObj* arg0)
                 continue;
             }
 
-            Player_80036978(i, (s32) &world_pos);
+            Player_80036978(i, &world_pos);
             is_outside = true;
             if (!(world_pos.x < Stage_GetCamBoundsLeftOffset()) &&
                 !(world_pos.x > Stage_GetCamBoundsRightOffset()))
